Writing JSON serialization and deserialization features for Jobs
It was necessary to rename the Attributes,
Defenses, and Pools in AttributeKeys to write
these features, as JSON does not natively support
Objects as keys. The workaround to this is to
only use enums that resolve to and from strings.

This meant that, for instance, STRENGTH = "STR"
would not work. And given that the abbreviated
form is already being used in the HTML, that was
the one chosen to be used in both places.

Also, to make serialization work on the Set
values in Jobs, such as affectedAttributes, they
are serialized into the same format as the Maps
are, however, they are deserialized back into the
original format using specialized methods.